+++ R/read_excel_export.R
@@ -6,13 +6,14 @@ read_excel_allsheets <- function(filename, tibble = FALSE) {
   # but if you like tidyverse tibbles (the default with read_excel)
   # then just pass tibble = TRUE
   sheets <- readxl::excel_sheets(filename)
-  x <- lapply(sheets,
-              function(X) {
-                readxl::read_excel(filename,
-                                   sheet = X,
-                                   range = readxl::cell_limits(c(15,1),c(NA,8)))}
-  )
-  if(!tibble) x <- lapply(x, as.data.frame)
+  x <- lapply(sheets, function(X) {
+    readxl::read_excel(
+      filename,
+      sheet = X,
+      range = readxl::cell_limits(c(15, 1), c(NA, 8))
+    )
+  })
+  if (!tibble) x <- lapply(x, as.data.frame)
   names(x) <- sheets
   x
 }
@@ -25,13 +26,14 @@ read_excel_datasheets <- function(filename, tibble = FALSE) {
   sheets <- readxl::excel_sheets(filename)
   # remove summary from sheets
   sheets <- stringr::str_subset(sheets, pattern = "Summary", negate = TRUE)
-  x <- lapply(sheets,
-              function(X) {
-                readxl::read_excel(filename,
-                                   sheet = X,
-                                   range = readxl::cell_limits(c(15,1),c(NA,8)))}
-  )
-  if(!tibble) x <- lapply(x, as.data.frame)
+  x <- lapply(sheets, function(X) {
+    readxl::read_excel(
+      filename,
+      sheet = X,
+      range = readxl::cell_limits(c(15, 1), c(NA, 8))
+    )
+  })
+  if (!tibble) x <- lapply(x, as.data.frame)
   names(x) <- sheets
   x
 }
@@ -43,11 +45,11 @@ read_excel_datasheets <- function(filename, tibble = FALSE) {
 flatten_with_session_names <- function(nested_list) {
   # Adding session names to each dataframe
   # TODO: defensive here, check for named list input
-  named_list <- imap(nested_list, ~ map(.x, mutate, session_name = .y))
+  named_list <- purrr::imap(nested_list, ~ map(.x, mutate, session_name = .y))
   # Flattening the list of lists to a single list of data frames
-  flat_list <- flatten(named_list)
+  flat_list <- purrr::list_flatten(named_list)
   # Binding all data frames into one
-  combined_df <- bind_rows(flat_list)
+  combined_df <- dplyr::bind_rows(flat_list)
   return(combined_df)
 }
 
@@ -58,14 +60,18 @@ extract_uid_dates <- function(filepaths) {
   # Regular expression to match the pattern
   pattern <- "\\d{4}_\\d{2}_\\d{2}_\\d{2}_\\d{2}_\\d{2}_thru_\\d{4}_\\d{2}_\\d{2}_\\d{2}_\\d{2}_\\d{2}"
   # Extract matches using stringr's str_extract
-  dates <- str_extract(filepaths, pattern)
+  dates <- stringr::str_extract(filepaths, pattern)
   return(dates)
 }
 
 parse_uid_session_dt <- function(df) {
   df |>
-    separate(session_name, into = c("session_start", "session_end"), sep = "_thru_") %>%
-    mutate(
+    tidyr::separate(
+      session_name,
+      into = c("session_start", "session_end"),
+      sep = "_thru_"
+    ) |>
+    dplyr::mutate(
       session_start = lubridate::ymd_hms(session_start),
       session_end = lubridate::ymd_hms(session_end)
     )