+++ NEWS.md
@@ -1,3 +1,4 @@
 # uid (development version)
 
 * Initial Github submission.
+* Adding basic function to read, write, process